backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.0
httpx==0.24.1
asyncpg>=0.29.0
pytesseract==0.3.10
pdfplumber==0.7.6
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow>=10.2.0
python-multipart==0.0.6
